requirements.txt
langchain
langchain_pinecone
langchain_openai
langchain_community
langchain_core
pinecone-client
python-dotenv
langgraph
jq
langchain-google-genai
pandas
langchain_anthropic
langchain_groq
gradio
pydantic>=2.0.0
python-telegram-bot
google-auth-oauthlib
google-auth-httplib2
google-api-python-client